requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
cmake==3.31.6
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
numpy==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
setuptools==75.8.2
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.45.1
